sui_analytics_indexer::handlers

Module tables

Table-specific processors for analytics pipelines.

Each module implements a Processor that extracts data from checkpoints and produces typed entries for a specific analytics table.

Re-exports

pub use checkpoint::CheckpointProcessor;

pub use df::DynamicFieldProcessor;

pub use event::EventProcessor;

pub use move_call::MoveCallProcessor;

pub use object::ObjectProcessor;

pub use package::PackageProcessor;

pub use package_bcs::PackageBCSProcessor;

pub use transaction::TransactionProcessor;

pub use transaction_bcs::TransactionBCSProcessor;

pub use transaction_objects::TransactionObjectsProcessor;

pub use wrapped_object::WrappedObjectProcessor;

Modules

checkpoint

df

event

move_call

object

package

package_bcs

transaction

transaction_bcs

transaction_objects

wrapped_object

Structs

InputObjectTracker

ObjectStatusTracker

WrappedStruct

Functions

get_is_consensus

get_move_struct

get_owner_address

get_owner_type

initial_shared_version

parse_struct